removing button component and applying styles to child button elements
