+++ packages/components/package.json
@@ -155,7 +155,6 @@
       "./components/hds/code-block/description.js": "./dist/_app_/components/hds/code-block/description.js",
       "./components/hds/code-block/index.js": "./dist/_app_/components/hds/code-block/index.js",
       "./components/hds/code-block/title.js": "./dist/_app_/components/hds/code-block/title.js",
-      "./components/hds/code-editor/button.js": "./dist/_app_/components/hds/code-editor/button.js",
       "./components/hds/code-editor/description.js": "./dist/_app_/components/hds/code-editor/description.js",
       "./components/hds/code-editor/full-screen-button.js": "./dist/_app_/components/hds/code-editor/full-screen-button.js",
       "./components/hds/code-editor/index.js": "./dist/_app_/components/hds/code-editor/index.js",
