--- packages/components/src/components/hds/code-editor/button.hbs
@@ -1,8 +1,0 @@
-<Hds::Button
-  @text={{@text}}
-  @size="small"
-  @icon={{@icon}}
-  @isIconOnly={{@isIconOnly}}
-  class="hds-code-editor__button"
-  ...attributes
-/>
