--- packages/components/src/components/hds/code-editor/button.ts
@@ -1,11 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: MPL-2.0
- */
-
-import TemplateOnlyComponent from '@ember/component/template-only';
-import type { HdsButtonSignature } from '../button';
-
-const HdsCodeEditorButton = TemplateOnlyComponent<HdsButtonSignature>();
-
-export default HdsCodeEditorButton;
