+++ packages/components/src/components/hds/code-editor/description.hbs
@@ -1,8 +1,3 @@
-<Hds::Text::Body
-  class="hds-code-editor__description hds-code-editor__header-content-text"
-  @tag="p"
-  @size="100"
-  ...attributes
->
+<Hds::Text::Body class="hds-code-editor__description" @tag="p" @size="100" ...attributes>
   {{yield}}
 </Hds::Text::Body>
