+++ packages/components/src/components/hds/code-editor/title.hbs
@@ -1,9 +1,3 @@
-<Hds::Text::Body
-  class="hds-code-editor__title hds-code-editor__header-content-text"
-  @tag={{this.tag}}
-  @size="200"
-  @weight="semibold"
-  ...attributes
->
+<Hds::Text::Body class="hds-code-editor__title" @tag={{this.tag}} @size="200" @weight="semibold" ...attributes>
   {{yield}}
 </Hds::Text::Body>
