+++ showcase/app/templates/components/code-editor.hbs
@@ -56,8 +56,8 @@
   <Shw::Text::H3>Custom content</Shw::Text::H3>
   <Shw::Grid @columns={{2}} @gap="2rem" as |SG|>
     <SG.Item @label="With custom header content">
-      <Hds::CodeEditor as |CE|>
-        <CE.Button @text="Custom button" />
+      <Hds::CodeEditor>
+        <Hds::Button @text="Custom button" />
       </Hds::CodeEditor>
     </SG.Item>
   </Shw::Grid>
@@ -74,21 +74,15 @@
       <Hds::CodeEditor as |CE|>
         <CE.Title>Code editor with title</CE.Title>
         <CE.Description>This is a code editor with a description</CE.Description>
-        <CE.Button @text="Custom button" />
+        <Hds::Button @text="Custom button" />
         <Hds::Icon @name="code" @color="#fff" />
       </Hds::CodeEditor>
     </SG.Item>
-    <SG.Item @label="With title, description, and internal and custom content">
+    <SG.Item @label="With title, description, internal actions, and custom content">
       <Hds::CodeEditor @hasFullScreenButton={{true}} @hasCopyButton={{true}} as |CE|>
         <CE.Title>Code editor with title</CE.Title>
         <CE.Description>This is a code editor with a description</CE.Description>
-        <CE.Button @text="Custom button" />
-        <Hds::Icon @name="code" @color="#fff" />
-      </Hds::CodeEditor>
-    </SG.Item>
-    <SG.Item @label="With internal and custom content">
-      <Hds::CodeEditor @hasFullScreenButton={{true}} @hasCopyButton={{true}} as |CE|>
-        <CE.Button @text="Custom button" />
+        <Hds::Button @text="Custom button" />
         <Hds::Icon @name="code" @color="#fff" />
       </Hds::CodeEditor>
     </SG.Item>
